backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pillow
pytesseract
pymupdf
google-generativeai
python-dotenv
